refactor: split REST resource routing template
