+++ .sampo/changesets/1018-split-rest-resource-routing-template.md
@@ -1,0 +1,8 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split generated REST resource PHP routing and controller template helpers from
+the main REST resource PHP template entrypoint while preserving generated
+output compatibility.
